slam-rs: move clip dumping into API and use core calibration
