+++ packages/slam-rs/tests/tools/dump_clip.py
@@ -1,271 +1,8 @@
-"""Dump catalog inputs for optional offline Rust replay and isolated timing.
+"""Dump catalog inputs for offline replay and timing."""
 
-The directory holds clip metadata, calibration JSON, IMU CSV/JSON, frame hashes,
-gray8 PGMs and integer timestamps. The optional NPZ and calibration pickle
-feed bench_track without decoding in its timing loop. NPZ holds frames in
-memory, so bound benchmark dumps with max_framesets.
-
-Catalog calibration uses f32 statics; fixture calibration uses the checked-in
-JSON doubles. Keeping this explicit permits sensitivity measurements.
-A full MIO07 stereo dump is about 7.5 GB and is not committed.
-"""
-
-import hashlib
-import json
-import pickle
-from dataclasses import dataclass
-from pathlib import Path
-from typing import Any, Literal
-
-import numpy as np
 import tyro
-from jaxtyping import Float64, Int64, UInt8
-from numpy import ndarray
 
-from slam_rs.catalog_feed import CameraCalib, CatalogSegment, Frameset, SegmentFeed, open_segment
-from slam_rs.reference import ReferenceManifest, ReferenceSegment, load_manifest, resolved_flow_config
-
-FIXTURES: Path = Path(__file__).resolve().parents[2] / "crates/slam-rs/tests/fixtures"
-"""Directory holding MSD calibration JSON files."""
-
-DEVICE_CALIBRATION: dict[str, str] = {"msd-index": "msdmi_calib.json", "msd-g2": "msdmg_calib.json"}
-"""Catalog dataset to its calibration JSON with double-precision values."""
-
-
-@dataclass(slots=True)
-class Config:
-    """Dump one reference segment's pixels, inertial samples and calibration."""
-
-    segment: str
-    """Segment id from ``reference_segments.toml``."""
-    output: Path
-    """Directory the clip is written to; created if missing."""
-    calibration: Literal["catalog", "fixture"] = "catalog"
-    """``catalog`` writes float32 statics; ``fixture`` writes calibration-file doubles."""
-    max_framesets: int | None = None
-    """Stop after this many framesets; None dumps the whole segment."""
-    npz: bool = False
-    """Also write ``clip.npz`` and ``clip.npz.calib.pkl``, which is what ``bench_track`` replays."""
-    window_s: float = 60.0
-    """Longest time window fetched in one round trip."""
-
-
-def eigen_quaternion_xyzw(rotation: Float64[ndarray, "3 3"]) -> Float64[ndarray, " 4"]:
-    """Convert a rotation matrix using trace and largest-diagonal quaternion branches.
-
-    Catalog matrices are stored in f32 and can be slightly non-orthogonal.
-    The explicit conversion keeps deterministic rounding without normalization.
-
-    Args:
-        rotation: Rotation block of ``imu_T_cam``.
-
-    Returns:
-        The quaternion as ``[qx, qy, qz, qw]``, without normalization.
-    """
-    quaternion: Float64[ndarray, " 4"] = np.zeros(4, dtype=np.float64)
-    trace: float = float(rotation.trace())
-    if trace > 0.0:
-        root: float = float(np.sqrt(trace + 1.0))
-        quaternion[3] = 0.5 * root
-        root = 0.5 / root
-        quaternion[0] = (rotation[2, 1] - rotation[1, 2]) * root
-        quaternion[1] = (rotation[0, 2] - rotation[2, 0]) * root
-        quaternion[2] = (rotation[1, 0] - rotation[0, 1]) * root
-        return quaternion
-    i: int = 0
-    if rotation[1, 1] > rotation[0, 0]:
-        i = 1
-    if rotation[2, 2] > rotation[i, i]:
-        i = 2
-    j: int = (i + 1) % 3
-    k: int = (j + 1) % 3
-    root = float(np.sqrt(rotation[i, i] - rotation[j, j] - rotation[k, k] + 1.0))
-    quaternion[i] = 0.5 * root
-    root = 0.5 / root
-    quaternion[3] = (rotation[k, j] - rotation[j, k]) * root
-    quaternion[j] = (rotation[j, i] + rotation[i, j]) * root
-    quaternion[k] = (rotation[k, i] + rotation[i, k]) * root
-    return quaternion
-
-
-def catalog_calibration(feed: SegmentFeed, fork_file: dict[str, Any]) -> dict[str, Any]:
-    """Build calibration JSON using catalog intrinsics, resolution and extrinsics.
-
-    Noise, update rate and time offset remain from the calibration file because
-    the recording does not carry them.
-
-    Args:
-        feed: Open segment feed, whose cameras carry the catalog geometry.
-        fork_file: The calibration JSON, already unwrapped from ``value0``.
-
-    Returns:
-        The ``value0`` body of a calibration JSON file.
-    """
-    calibration: dict[str, Any] = dict(fork_file)
-    poses: list[dict[str, float]] = []
-    intrinsics: list[dict[str, Any]] = []
-    resolution: list[list[int]] = []
-    camera: CameraCalib
-    for camera in feed.cameras:
-        quaternion: Float64[ndarray, " 4"] = eigen_quaternion_xyzw(camera.imu_T_cam[:3, :3])
-        poses.append(
-            {
-                "px": float(camera.imu_T_cam[0, 3]),
-                "py": float(camera.imu_T_cam[1, 3]),
-                "pz": float(camera.imu_T_cam[2, 3]),
-                "qx": float(quaternion[0]),
-                "qy": float(quaternion[1]),
-                "qz": float(quaternion[2]),
-                "qw": float(quaternion[3]),
-            }
-        )
-        terms: dict[str, float] = {"fx": camera.fx, "fy": camera.fy, "cx": camera.cx, "cy": camera.cy}
-        if camera.model == "kb4":
-            terms |= dict(zip(["k1", "k2", "k3", "k4"], camera.distortion.tolist(), strict=True))
-            intrinsics.append({"camera_type": "kb4", "intrinsics": terms})
-        else:
-            terms |= dict(zip(["k1", "k2", "p1", "p2", "k3", "k4", "k5", "k6"], camera.distortion.tolist(), strict=True))
-            if camera.distortion_valid_radius is None:
-                raise ValueError(f"cam_{camera.index:02d} is radtan8 without a valid radius; basalt needs rpmax")
-            terms["rpmax"] = camera.distortion_valid_radius
-            intrinsics.append({"camera_type": "pinhole-radtan8", "intrinsics": terms})
-        resolution.append([camera.width, camera.height])
-    calibration["T_imu_cam"] = poses
-    calibration["intrinsics"] = intrinsics
-    calibration["resolution"] = resolution
-    return calibration
-
-
-def write_pgm(path: Path, image: UInt8[ndarray, "h w"]) -> None:
-    """Write one grayscale image in the binary PGM the Rust lanes read.
-
-    Args:
-        path: Destination file.
-        image: C-contiguous 8-bit raster.
-    """
-    header: bytes = f"P5\n{image.shape[1]} {image.shape[0]}\n255\n".encode("ascii")
-    path.write_bytes(header + image.tobytes())
-
-
-def write_oracle_inputs(output: Path, frame_t_ns: list[int], imu_lines: list[str]) -> None:
-    """Write integer frameset timestamps and IMU JSON beside the PGMs.
-    These duplicate the clock and samples in clip.json and imu.csv for replay tools.
-
-    Args:
-        output: Clip directory.
-        frame_t_ns: Frameset timestamps, in order.
-        imu_lines: The data rows of ``imu.csv``, without the header.
-    """
-    (output / "timestamps.txt").write_text("".join(f"{t_ns}\n" for t_ns in frame_t_ns))
-    samples: list[str] = []
-    for line in imu_lines:
-        fields: list[str] = line.split(",")
-        samples.append(f'    {{"t_ns": {fields[0]}, "gyro": [{", ".join(fields[1:4])}], "accel": [{", ".join(fields[4:7])}]}}')
-    (output / "imu.json").write_text('{"imu": [\n' + ",\n".join(samples) + "\n]}\n")
-
-
-def main(config: Config) -> None:
-    """Dump the segment named by the config.
-
-    Args:
-        config: Parsed CLI options.
-
-    Raises:
-        ValueError: If ``--npz`` was asked for with no frameset to bundle, if the
-            segment is not in the manifest, or if its dataset has no fork
-            calibration.
-    """
-    # Before the manifest, the catalog and the output directory: zero framesets
-    # was accepted, every non-NPZ side file was written, and then `np.stack` on
-    # the empty image list raised out of NumPy — leaving a directory that reads
-    # as a clip and holds no frameset (S25 review).
-    if config.npz and config.max_framesets == 0:
-        raise ValueError("--max-framesets 0 with --npz has no frameset to bundle; the bench dump needs at least one")
-    manifest: ReferenceManifest = load_manifest()
-    segment: ReferenceSegment = manifest.by_id(config.segment)
-    if segment.dataset_name not in DEVICE_CALIBRATION:
-        raise ValueError(f"{segment.dataset_name} has no fork calibration file")
-    fork_file: dict[str, Any] = json.loads((FIXTURES / DEVICE_CALIBRATION[segment.dataset_name]).read_text())["value0"]
-
-    config.output.mkdir(parents=True, exist_ok=True)
-    bundle_images: list[UInt8[ndarray, "n_cameras h w"]] = []
-    bundle_imu_counts: list[int] = []
-    bundle_imu_t: list[Int64[ndarray, " n_samples"]] = []
-    bundle_imu_gyro: list[Float64[ndarray, "n_samples 3"]] = []
-    bundle_imu_accel: list[Float64[ndarray, "n_samples 3"]] = []
-    imu_lines: list[str] = ["#t_ns,gyro_x,gyro_y,gyro_z,accel_x,accel_y,accel_z"]
-    frame_lines: list[str] = ["# t_ns,cam_index,sha256 of the gray8 HxW pixels pushed to basalt"]
-    frame_t_ns: list[int] = []
-    dumped: int = 0
-
-    with open_segment(
-        CatalogSegment(manifest.catalog_url, segment.dataset_name, segment.segment_id),
-        segment.imu,
-        window_s=config.window_s,
-    ) as feed:
-        calibration: dict[str, Any] = catalog_calibration(feed, fork_file) if config.calibration == "catalog" else fork_file
-        (config.output / "calib.json").write_text(json.dumps({"value0": calibration}, indent=2) + "\n")
-        frameset: Frameset
-        for frameset in feed.framesets():
-            if config.max_framesets is not None and dumped >= config.max_framesets:
-                break
-            for index, image in enumerate(frameset.images):
-                write_pgm(config.output / f"frame_{dumped:03d}_cam{index}.pgm", image)
-            if config.npz:
-                bundle_images.append(np.stack(frameset.images))
-                bundle_imu_counts.append(len(frameset.imu))
-                bundle_imu_t.append(frameset.imu.t_ns)
-                bundle_imu_gyro.append(frameset.imu.gyro_rad_s)
-                bundle_imu_accel.append(frameset.imu.accel_m_s2)
-            absolute_ns: int = frameset.t_ns + feed.capture_start_time_ns
-            frame_lines.extend(f"{absolute_ns},{index},{digest}" for index, digest in enumerate(frameset.image_digests()))
-            for t_ns, gyro, accel in zip(frameset.imu.t_ns.tolist(), frameset.imu.gyro_rad_s.tolist(), frameset.imu.accel_m_s2.tolist(), strict=True):
-                imu_lines.append(f"{t_ns}," + ",".join(repr(value) for value in [*gyro, *accel]))
-            frame_t_ns.append(frameset.t_ns)
-            dumped += 1
-            if dumped % 200 == 0:
-                print(f"{dumped} framesets", flush=True)
-
-        clip: dict[str, Any] = {
-            "segment_id": feed.segment_id,
-            "dataset_name": segment.dataset_name,
-            "capture_start_time_ns": feed.capture_start_time_ns,
-            "calibration_source": config.calibration,
-            "num_cameras": len(feed.cameras),
-            "resolution_wh": [[camera.width, camera.height] for camera in feed.cameras],
-            "framesets": dumped,
-            "frame_t_ns": frame_t_ns,
-            "imu_samples": len(imu_lines) - 1,
-            "decode_path": segment.decode_path,
-        }
-    (config.output / "imu.csv").write_text("\n".join(imu_lines) + "\n")
-    write_oracle_inputs(config.output, frame_t_ns, imu_lines[1:])
-    (config.output / "frames.sha256").write_text("\n".join(frame_lines) + "\n")
-    (config.output / "clip.json").write_text(json.dumps(clip, indent=2) + "\n")
-    if config.npz:
-        bundle: Path = config.output / "clip.npz"
-        np.savez(
-            bundle,
-            images=np.stack(bundle_images),
-            t_ns=np.array(frame_t_ns, dtype=np.int64),
-            imu_counts=np.array(bundle_imu_counts, dtype=np.int64),
-            imu_t=np.concatenate(bundle_imu_t),
-            imu_g=np.concatenate(bundle_imu_gyro),
-            imu_a=np.concatenate(bundle_imu_accel),
-            safe_radius=np.int64(resolved_flow_config(manifest, segment)[0].optical_flow_image_safe_radius),
-        )
-        # The dataclasses the feed built, because `bench_track` compares lanes on
-        # the calibration the reference ran and not on a second derivation of it.
-        # Pickle rather than JSON for the same reason: these two values are the
-        # feed's own, and a hand-written schema here would be a third account of
-        # what a calibration is. Producer and consumer therefore ship together.
-        with bundle.with_suffix(bundle.suffix + ".calib.pkl").open("wb") as handle:
-            pickle.dump({"cameras": feed.cameras, "imu": feed.imu}, handle)
-        print(f"clip.npz {np.stack(bundle_images).shape} + calib.pkl -> {bundle}")
-    digest: str = hashlib.sha256((config.output / "frames.sha256").read_bytes()).hexdigest()
-    print(f"{dumped} framesets, {clip['imu_samples']} inertial samples -> {config.output}")
-    print(f"frames.sha256 {digest}")
-
+from slam_rs.apis.dump_clip import Config, main
 
 if __name__ == "__main__":
     main(tyro.cli(Config))
